requirements.txt
celery==4.0.2
Django==1.10.5
django-celery-results==1.0.1
djangorestframework==3.5.3
docker==2.0.2
flower==0.9.1
psycopg2==2.6.2
redis==2.10.5
requests==2.12.5
